requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
filelock==3.18.0
Flask==3.1.0
flask-ngrok==0.0.25
fsspec==2025.3.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
pillow==11.1.0
pyngrok==7.2.3
PyYAML==6.0.2
requests==2.32.3
sympy==1.13.1
torch==2.6.0
torchvision==0.21.0
typing_extensions==4.13.1
urllib3==2.3.0
Werkzeug==3.1.3
